Revert "Time to merge all this (Authors, header, expand, etc, idk)"
