+++ .github/workflows/deploy.yml
@@ -1,8 +1,5 @@
 name: Deploy
-on:
-  push:
-    branches:
-      - master
+on: [push]
 
 jobs:
   deploy:
@@ -27,6 +24,6 @@ jobs:
       - name: Upload to Deno Deploy
         uses: denoland/deployctl@v1
         with:
-          project: readcycle-front 
+          project: readcycle-front # TODO: replace with Deno Deploy project name
           entrypoint: server/entry.mjs
           root: dist
