+++ astro.config.mjs
@@ -3,8 +3,11 @@ import tailwind from "@astrojs/tailwind";
 import svelte from "@astrojs/svelte";
 import solidJs from "@astrojs/solid-js";
 
+import deno from "@astrojs/deno";
+
 // https://astro.build/config
 export default defineConfig({
   output: "server",
   integrations: [tailwind(), svelte(), solidJs()],
+  adapter: deno()
 });
