+++ src/components/Header.svelte
@@ -1,33 +1,17 @@
-<script>
-  import  pb from '../pocketbase.ts'
-
-  export let userData;
-  // N'est meme pas exécutée
-  const logout = async () => {
-    pb.authStore.clear()
-  }
-
-</script>
-
 <navbar> 
   <header class="navbar bg-primary mb-12">
     <div class="flex-1">
       <a class="btn btn-ghost normal-case text-x1" href="/">TBD</a>
-      <img src="/logo.svg" alt="logo"/>
-    </div>
-    <div class="grow">
-      <form method="POST">
-        <input type="search" placeholder="Chercher un livre, un auteur" class="input input-bordered input-secondary w-full max-w-xs" />
-      </form>
+      <img src="logo.svg" alt="logo"/>
     </div>
     <div class="flex-none">
       <div class="flex justify-end flex-1 px-2">
         <div class="flex items-stretch">
           <div class="flex place-items-center">
-            {#if userData.isValid}
-              <a class="link" href="/profile">{userData.model.name}</a>
+            {#if 1==1}
+              <a href="/login" class="link">Connection</a>
             {:else}
-              <a  class="link" href="/login">Connection</a>
+              <p class="inline-block">Connecté</p>
             {/if}
           </div>
           <div class="dropdown dropdown-end">
@@ -35,15 +19,9 @@
               <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" class="inline-block w-5 h-5 stroke-current"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M5 12h.01M12 12h.01M19 12h.01M6 12a1 1 0 11-2 0 1 1 0 012 0zm7 0a1 1 0 11-2 0 1 1 0 012 0zm7 0a1 1 0 11-2 0 1 1 0 012 0z"></path></svg> 
             </label>
             <ul tabindex="0" class="menu dropdown-content z-[1] p-2 shadow bg-base-200 rounded-box w-52 mt-4">
-              {#if userData.isValid}
-                <li><a class="link" href="/profile">Profile</a></li>
-                <!-- Ici on ne peut malheuresement pas accéder au propriétés d'astro, et donc je pense créer un page à part entière pour logout-->
-                <li><button onClick="" on:click={logout} class="link">Déconnexion</button></li>
-              {:else}
-                <li><a class="link" href="/login">Connection</a></li>
-              {/if}
-              <li><a class="link" href="/books">Livres</a></li>
-              <li><a class="link" href="/authors">Auteurs</a></li>
+              <li><a class="link" href="/profile">Profile</a></li> 
+              <li><button onclick="logout.showModal()" on:click="{handleSignOut}" class="link">Déconnexion</button></li>
+              <li><a class="link" href="/books">Books</a></li>
             </ul>
           </div>
         </div>
