+++ src/components/UserProfile.svelte
@@ -1,10 +1,5 @@
-<script lang="ts">
-   export let userData: any;
+<script>
+    export let userData;
 </script>
 
-{#if userData}
-  <h1>Profile de {userData.username}</h1>
-{:else}
-  <h1>Il semblmerait que vous ne soyez pas connecté ! </h1>
-  <a href='/login' class="link primary">Créez un compte ou connectez vous ici</a>
-{/if}
+<h1>Profile de {userData.username}</h1>
