+++ src/components/books.svelte
@@ -1,0 +1,14 @@
+<script>
+import BookItem from "./book-item.svelte";
+
+export let resultat
+
+</script>                     
+    
+<ul class="flex">
+    {#each resultat.items as book}
+        <li>
+            <BookItem Title={book.title} Author={book.author_id} Genre={book.genres_id} Editor={book.editor} ISBN={book.isbn} Summary={book.summary}/>
+        </li>
+    {/each}
+</ul>
