+++ src/layouts/Layout.astro
@@ -6,9 +6,6 @@ interface Props {
 }
 
 const { title } = Astro.props; 
-
-const locals = Astro.locals;
-const data = locals.pb.authStore;
 ---
 
 <!DOCTYPE html>
@@ -17,13 +14,13 @@ const data = locals.pb.authStore;
 		<meta charset="UTF-8" />
 		<meta name="description" content="Astro description">
 		<meta name="viewport" content="width=device-width" />
-		<link rel="icon" type="image/svg+xml" href="/logo.svg" />
+		<link rel="icon" type="image/svg+xml" href="/favicon.svg" />
 		<meta name="generator" content={Astro.generator} />
 		<title>{title}</title>
 		<ViewTransitions />
 		</head>
 	<body class="bg-base-1OO">
-    <Header userData={data}/>
+    <Header />
 
 		<main class="flex flex-col w-full place-items-center">
 			<slot/>
