--- src/pages/authors.astro
@@ -1,32 +1,0 @@
----
-import Layout from "../layouts/Layout.astro";
-
-const locals = Astro.locals;
-
-const result = await locals.pb.collection("authors").getList(1, 2);
-
----
-
-<Layout title="Authors - Readcyle">
-  <h1 class="inline-block text-3xl sm:text-3xl font-extrabold text-primary-content tracking-tigh">
-    Ca fait beaucoup de noms inconnus !
-  </h1>
-  <h2 class="italic text-accent-content">
-    Ils en ont fait des grandes choses pourtant
-  </h2>
-  <ul class="flex">
-    {
-      result.items.map((author: any) => (
-        <li>
-          <a class="link" href=`/authors/${author.id}`>{author.first_name}, {author.last_name}</a>
-        </li>
-      ))
-    }
-  </ul>
-  <!--Composant à réaliser prenant un nombre comme props en et proposant les pages alentours-->
-  <div class="tabs tabs-boxed">
-    <a class="tab tab-active">Tab 1</a> 
-    <a class="tab">Tab 2</a> 
-    <a class="tab">Tab 3</a>
-  </div>
-</Layout>
