--- src/pages/authors/[author_id].astro
@@ -1,24 +1,0 @@
----
-import Layout from "../../layouts/Layout.astro";
-const { author_id } = Astro.params;
-
-const locals = Astro.locals;
-
-const result = await locals.pb.collection("authors").getOne(author_id);
-
-const title = `${result.first_name}, ${result.last_name}`
----
-<Layout title={title}>
-    <div class="hero min-h-screen bg-base-200">
-        <div class="hero-content flex-col lg:flex-row-reverse">
-          <img src="" alt={title} class="max-w-sm rounded-lg shadow-2xl" />
-          <div>
-            <h1 class="text-5xl font-bold">{title}</h1>
-            <!--On va en laisser à martin quand même -->
-            <p class="py-6">Provident cupiditate voluptatem et in. Quaerat fugiat ut assumenda excepturi exercitationem quasi. In deleniti eaque aut repudiandae et a id nisi.</p>
-            <button class="btn btn-primary">Get Started</button>
-          </div>
-        </div>
-      </div>
-</Layout>
-
