+++ src/pages/login.astro
@@ -61,13 +61,7 @@ const errors = { username: "", password: ""};
     }
     <input required name="password" type="password" placeholder="Mot de passe"  class="input input-bordered input-secondary w-full max-w-xs mb-8" />
     <br>
-    <button class="btn btn-primary mb-4" type="submit">
-      Connection
-       <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-6 h-6">
-        <path stroke-linecap="round" stroke-linejoin="round" d="M11.25 4.5l7.5 7.5-7.5 7.5m-6-15l7.5 7.5-7.5 7.5" />
-      </svg>
-
-    </button>
+    <button class="btn btn-primary mb-4" type="submit">Connection</button>
   </form>
 
   <p>Vous n'avez pas de compte ? <a href="/signup" class="link link-focus">Cliquez ici pour en créer un</a></p>
