+++ src/pages/profile.astro
@@ -1,7 +1,6 @@
 ---
 import Layout from "../layouts/Layout.astro"
 import UserProfile from "../components/UserProfile.svelte";
-
  const locals = Astro.locals;
  const data = locals.pb.authStore.model;
 ---
