+++ src/pages/signup.astro
@@ -44,7 +44,7 @@ const errors = { email: "", password: ""};
 ---
 
 <Layout title='Signup - ReadCycle'>
-<h1 class="text-base-content text-2xl mb-4 ">Créer un compte</h1>
+<h1 class="text-base-content text-2xl mb-4 ">Connection</h1>
     <form method="post" class="place-items-center flex-col flex w-full">
     {
         errors.email && 
@@ -64,12 +64,8 @@ const errors = { email: "", password: ""};
     }
     <input required name="password" type="password" placeholder="Mot de passe"  class="input input-bordered input-secondary w-full max-w-xs mb-8" />
     <br>
-    <button class="btn btn-primary mb-4" type="submit">
-      Continuer
-      <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-6 h-6">
-        <path stroke-linecap="round" stroke-linejoin="round" d="M11.25 4.5l7.5 7.5-7.5 7.5m-6-15l7.5 7.5-7.5 7.5" />
-      </svg>
-    </button>
+    <button class="btn btn-primary mb-4" type="submit">Connection</button>
   </form>
 
+  <p>Vous n'avez pas de compte ? <a href="/signup" class="link link-focus">Cliquez ici pour en créer un</a></p> 
 </Layout>
